added switch case to get a single command
